@@ -20,7 +20,7 @@
     }
   ],
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1",
     "tooltipster/tooltipster": "4.2.*",
     "ext-mysqli": "*",
     "ext-json": "*",
