@@ -15,16 +15,16 @@ namespace Ogsteam\Ogspy\Model;
 
 use Ogsteam\Ogspy\Abstracts\Model_Abstract;
 
-class Config_Model  extends Model_Abstract
+class Config_Model extends Model_Abstract
 {
-
+    //TODO: There is no method to create a new configuration
     /**
      * Retourne tous les elements de la configuration
      * @return array
      */
     public function get_all()
     {
-        $output = array();
+        $output = [];
         $request = "SELECT * from " . TABLE_CONFIG;
         $result = $this->db->sql_query($request);
         // Output config as PHP code
@@ -32,6 +32,24 @@ class Config_Model  extends Model_Abstract
             $output[$cur_config_item[0]] = stripslashes($cur_config_item[1]);
         }
         return $output;
+    }
+
+    /**
+     * Fonction de recherche de la configuration
+     * @param array|null $filter
+     * @return array|bool
+     */
+    public function get(array $filter = null): bool|array|null
+    {
+        if ($filter === null) {
+            return false;
+        }
+
+        $queryStr = "'" . implode("','", $filter) . "'";
+        $query = "SELECT `config_name`, `config_value` FROM " . TABLE_CONFIG . " WHERE `config_name` IN ($queryStr)";
+
+        $result = $this->db->sql_query($query);
+        return $this->db->sql_fetch_assoc($result);
     }
 
     /**
@@ -44,6 +62,7 @@ class Config_Model  extends Model_Abstract
         if ($filter == null) {
             $filter = array();
         }
+
         $query = "SELECT `config_name`, `config_value` FROM " . TABLE_CONFIG;
         $i = 0;
         foreach ($filter as $key => $value) {
@@ -79,14 +98,15 @@ class Config_Model  extends Model_Abstract
      * Met à jour la config
      * @$config_value valeur de la configuration
      * @$config_name nom de la configuration
-     * @param $config_value
-     * @param $config_name
+     * @param $configName
+     * @param $configValue
      */
-    public function update_one($config_value, $config_name)
+    public function update_one( $configValue, $configName)
     {
-        $query = "UPDATE " . TABLE_CONFIG . " SET
-                    `config_value` = '" . $this->db->sql_escape_string($config_value) . "'
-                 WHERE `config_name` = '" . $this->db->sql_escape_string($config_name) . "'";
+        $query = "INSERT INTO " . TABLE_CONFIG . " (`config_name`, `config_value`)
+              VALUES ('" . $this->db->sql_escape_string($configName) . "',
+                      '" . $this->db->sql_escape_string($configValue) . "')
+              ON DUPLICATE KEY UPDATE `config_value` = '" . $this->db->sql_escape_string($configValue) . "'";
         $this->db->sql_query($query);
     }
 }
